+++ client/package.json
@@ -6,6 +6,7 @@
     "firebase": "^6.3.0",
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
+    "react-firestore": "^1.0.1",
     "react-router-dom": "^5.0.1",
     "react-scripts": "2.1.1",
     "rimble-ui": "^0.9.4",
