+++ client/src/App.js
@@ -6,18 +6,27 @@ import { Header , Body } from './components'
 
 import './styles/core.css';
 
+import firebase from '@firebase/app';
+import '@firebase/firestore';
+import { FirestoreProvider } from 'react-firestore';
+import firebaseConfig from './firebase-config';
+
+firebase.initializeApp(firebaseConfig);
+
 const App = ()=><>
     <Header/>
-    <Body>
-      <Switch>
-        {routes.map(route =>
-          <Route
-            {...route}
-            key={route.path}
-          />
-        )}
-      </Switch>
-    </Body>
+    <FirestoreProvider firebase={firebase}>
+      <Body>
+        <Switch>
+          {routes.map(route =>
+            <Route
+              {...route}
+              key={route.path}
+            />
+          )}
+        </Switch>
+      </Body>
+    </FirestoreProvider>
 </>
 
 export default withRouter(App);
