+++ client/src/components/ItemCard/index.js
@@ -3,11 +3,11 @@ import { Link } from 'react-router-dom';
 import Web3  from 'web3';
 import { Image, Card, Heading, Text} from 'rimble-ui';
 
-const ItemCard = ({ id, name , image, price}) => <Link to={`/item/${id}`}>
+const ItemCard = ({ id, name , image, priceInWei}) => <Link to={`/item/${id}`}>
     <Card>
         <Heading.h4>{name}</Heading.h4>
         <Image borderRadius={8} src={image}/>
-        <Text bold>{Web3.utils.fromWei(price, 'ether')} ETH</Text>
+        <Text bold>{Web3.utils.fromWei(priceInWei, 'ether')} ETH</Text>
     </Card></Link>
 
 export default ItemCard;
