+++ client/src/firebase-config.js
@@ -1,0 +1,11 @@
+var config = {
+    apiKey: "AIzaSyAr63kLwJ1R36cLW5RuNVdXZDXUa_8FatE",
+    authDomain: "saleplace-cbcc0612.firebaseapp.com",
+    databaseURL: "https://saleplace-cbcc0612.firebaseio.com",
+    projectId: "saleplace-cbcc0612",
+    storageBucket: "saleplace-cbcc0612.appspot.com",
+    messagingSenderId: "1008043798181",
+    appId: "1:1008043798181:web:7cecb342cd5d20fd"
+};
+
+export default config;
