+++ client/src/pages/ItemDetails/styledComponent.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const ContentFlex = styled.div`
+    display: flex;
+    flex: 1;
+    justify-content: space-between;
+    flex-basis: auto;
+    flex-direction: column;
+`
+
+export { ContentFlex}
